@@ -1,24 +1,22 @@
 name: ada_chat_flutter
 description: Flutter implementation of Ada chat
-version: 0.0.4
+version: 1.0.0
 
 environment:
-  sdk: '>=3.0.0 <4.0.0'
+  sdk: '>=3.3.0 <4.0.0'
   flutter: ">=3.19.0"
 
 dependencies:
-  fk_user_agent: ^2.1.0 # Used by MockWebViewDependencies() tests helper
   flutter:
     sdk: flutter
-  flutter_inappwebview: ^6.0.0
   meta: ^1.8.0
-  mocktail: ^1.0.3 # Used by MockWebViewDependencies() tests helper
-  plugin_platform_interface: ^2.1.8 # Used by MockWebViewDependencies() tests helper
+  webview_flutter: ^4.8.0
 
 dev_dependencies:
-  flutter_lints: ^2.0.0
+  flutter_lints: ^4.0.0
   flutter_test:
     sdk: flutter
+  mocktail: ^1.0.3
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
@@ -30,9 +28,6 @@ flutter:
   # assets:
   #   - images/a_dot_burr.jpeg
   #   - images/a_dot_ham.jpeg
-  assets:
-    - assets/
-
   #
   # For details regarding assets in packages, see
   # https://flutter.dev/assets-and-images/#from-packages
