@@ -1,5 +1,6 @@
 import 'dart:async';
 import 'dart:convert';
+import 'dart:developer';
 
 import 'package:ada_chat_flutter/src/ada_controller.dart';
 import 'package:ada_chat_flutter/src/ada_controller_init.dart';
@@ -8,7 +9,11 @@ import 'package:ada_chat_flutter/src/customized_web_view.dart';
 import 'package:ada_chat_flutter/src/utils.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_inappwebview/flutter_inappwebview.dart';
+import 'package:webview_flutter/webview_flutter.dart';
+// ignore: depend_on_referenced_packages
+import 'package:webview_flutter_android/webview_flutter_android.dart';
+// ignore: depend_on_referenced_packages
+import 'package:webview_flutter_wkwebview/webview_flutter_wkwebview.dart';
 
 /// Ada chat WebView widget.
 ///
@@ -16,11 +21,11 @@ import 'package:flutter_inappwebview/flutter_inappwebview.dart';
 class AdaWebView extends StatefulWidget {
   const AdaWebView({
     super.key,
+    required this.embedUri,
     required this.handle,
     this.name,
     this.email,
     this.phone,
-    this.urlRequest,
     this.controller,
     this.language = 'en',
     this.cluster,
@@ -37,8 +42,8 @@ class AdaWebView extends StatefulWidget {
     this.testMode = false,
     this.onProgressChanged,
     this.browserSettings,
-    this.onLoaded,
     this.onAdaReady,
+    this.onLoaded,
     this.onEvent,
     this.onConversationEnd,
     this.onDrawerToggle,
@@ -57,9 +62,9 @@ class AdaWebView extends StatefulWidget {
   /// User phone, an additional field to add to the metadata
   final String? phone;
 
-  /// Url to your own page with assets/embed.html file.
+  /// Url to your own page with embed.html file.
   /// The domain must be added to approved domains. (Doc)[https://docs.ada.cx/docs/scripted/use-ada-with-your-website/deploy-ada-on-your-website-or-app/restrict-your-bot-to-a-list-of-approved-domains/].
-  final URLRequest? urlRequest;
+  final Uri embedUri;
 
   /// Controller for Ada chat. The default implementation is [AdaController].
   final AdaControllerInit? controller;
@@ -90,9 +95,9 @@ class AdaWebView extends StatefulWidget {
   final void Function(dynamic isRolledOut)? onAdaReady;
   final void Function(dynamic event)? onEvent;
   final void Function(dynamic event)? onConversationEnd;
-  final void Function(bool isDrawerOpen)? onDrawerToggle;
+  final void Function(dynamic isDrawerOpen)? onDrawerToggle;
   final void Function(String level, String message)? onConsoleMessage;
-  final void Function(String request, String error)? onLoadingError;
+  final void Function(String request, String response)? onLoadingError;
 
   @override
   State<AdaWebView> createState() => _AdaWebViewState();
@@ -104,7 +109,7 @@ class AdaWebView extends StatefulWidget {
     properties.add(StringProperty('name', name));
     properties.add(StringProperty('email', email));
     properties.add(StringProperty('phone', phone));
-    properties.add(DiagnosticsProperty<URLRequest?>('urlRequest', urlRequest));
+    properties.add(DiagnosticsProperty<Uri?>('urlRequest', embedUri));
     properties.add(StringProperty('language', language));
     properties.add(StringProperty('cluster', cluster));
     properties.add(StringProperty('domain', domain));
@@ -133,56 +138,126 @@ class AdaWebView extends StatefulWidget {
 }
 
 class _AdaWebViewState extends State<AdaWebView> {
-  late final _settings = InAppWebViewSettings(
-    supportZoom: false,
-    horizontalScrollBarEnabled: false,
-    verticalScrollBarEnabled: false,
-    useWideViewPort: false,
-    disableDefaultErrorPage: true,
-    allowFileAccessFromFileURLs: _allowFileAccessFromFileURLs,
-    allowsBackForwardNavigationGestures: false,
-    disableContextMenu: true,
-  );
-
-  /// Unsafe feature. Needed if the embed.html file is not hosted anywhere, then
-  /// the file from the assets will be used.
-  /// More about it (here)[https://inappwebview.dev/docs/webview/in-app-webview#antipatterns].
-  late final bool _allowFileAccessFromFileURLs = widget.urlRequest == null;
+  late final WebViewController _controller;
 
   @override
-  Widget build(BuildContext context) => InAppWebView(
-        initialUrlRequest: widget.urlRequest,
-        initialFile: _getInitialFile,
-        initialSettings: _settings,
-        shouldOverrideUrlLoading: _shouldOverrideUrlLoading,
-        onProgressChanged: (_, progress) =>
-            widget.onProgressChanged?.call(progress),
-        onReceivedError: (controller, request, error) =>
-            widget.onLoadingError?.call(
-          request.toString(),
-          error.toString(),
-        ),
-        onReceivedHttpError: (controller, request, errorResponse) =>
-            widget.onLoadingError?.call(
-          request.toString(),
-          errorResponse.toString(),
-        ),
-        onWebViewCreated: _init,
-        onLoadStop: _start,
-        onConsoleMessage: (controller, consoleMessage) =>
-            widget.onConsoleMessage?.call(
-          consoleMessage.messageLevel.toString(),
-          consoleMessage.message,
+  void initState() {
+    super.initState();
+
+    late final PlatformWebViewControllerCreationParams params;
+
+    if (WebViewPlatform.instance is WebKitWebViewPlatform) {
+      params = WebKitWebViewControllerCreationParams(
+        allowsInlineMediaPlayback: true,
+        mediaTypesRequiringUserAction: const <PlaybackMediaTypes>{},
+      );
+    } else if (WebViewPlatform.instance is AndroidWebViewPlatform) {
+      params = AndroidWebViewControllerCreationParams();
+    } else {
+      params = const PlatformWebViewControllerCreationParams();
+    }
+
+    _controller = WebViewController.fromPlatformCreationParams(params);
+
+    final platform = _controller.platform;
+    if (platform is AndroidWebViewController) {
+      AndroidWebViewController.enableDebugging(kDebugMode);
+    }
+
+    _controller
+      ..setJavaScriptMode(JavaScriptMode.unrestricted)
+      ..setBackgroundColor(Colors.white)
+      ..setOnConsoleMessage(_onConsoleMessage)
+      ..enableZoom(false)
+      ..setNavigationDelegate(
+        NavigationDelegate(
+          onHttpAuthRequest: _onHttpAuthRequest,
+          onUrlChange: _onUrlChange,
+          onProgress: _onProgress,
+          onPageStarted: _onPageStarted,
+          onPageFinished: _onPageFinished,
+          onHttpError: _onHttpError,
+          onWebResourceError: _onWebResourceError,
+          onNavigationRequest: _onNavigationRequest,
         ),
       );
 
-  String? get _getInitialFile {
-    return widget.urlRequest != null
-        ? null
-        : 'packages/ada_chat_flutter/assets/embed.html';
+    _controller.loadRequest(widget.embedUri);
   }
 
-  Future<void> _start(InAppWebViewController controller, WebUri? url) async {
+  void _onConsoleMessage(JavaScriptConsoleMessage message) =>
+      widget.onConsoleMessage?.call(message.level.toString(), message.message);
+
+  FutureOr<NavigationDecision> _onNavigationRequest(NavigationRequest request) {
+    final uri = Uri.parse(request.url);
+    log('AdaWebView:onNavigationRequest: '
+        'url=${uri.toString()}, isMainFrame=${request.isMainFrame}');
+
+    if (_isAdaChatLink(uri) || _isAdaSupportLink(uri) || _isBlankPage(uri)) {
+      return NavigationDecision.navigate;
+    }
+
+    unawaited(
+      showAdaptiveDialog(
+        context: context,
+        barrierColor: Colors.transparent,
+        builder: (context) => CustomizedWebView(
+          url: uri,
+          browserSettings: widget.browserSettings,
+        ),
+        useRootNavigator: false,
+      ),
+    );
+
+    return NavigationDecision.prevent;
+  }
+
+  bool _isBlankPage(Uri uri) => uri.toString() == 'about:blank';
+
+  bool _isAdaSupportLink(Uri uri) => uri.host == '${widget.handle}.ada.support';
+
+  bool _isAdaChatLink(Uri uri) => uri == widget.embedUri;
+
+  void _onWebResourceError(WebResourceError error) {
+    log('AdaWebView:onWebResourceError: '
+        'errorCode=${error.errorCode}, '
+        'description=${error.description}');
+  }
+
+  void _onHttpError(HttpResponseError error) => widget.onLoadingError?.call(
+        error.request?.uri.toString() ?? '',
+        'uri=${error.response?.uri}, '
+        'statusCode=${error.response?.statusCode}, '
+        'headers=${error.response?.headers}, ',
+      );
+
+  void _onPageFinished(String url) {
+    log('AdaWebView:onPageFinished: url=$url');
+
+    Future.delayed(Duration.zero, () async {
+      await _init();
+      await _start();
+    });
+  }
+
+  void _onPageStarted(String url) {
+    log('AdaWebView:onPageStarted: url=$url');
+  }
+
+  void _onProgress(int progress) => widget.onProgressChanged?.call(progress);
+
+  void _onUrlChange(UrlChange change) {
+    log('AdaWebView:onUrlChange: url=${change.url}');
+  }
+
+  void _onHttpAuthRequest(HttpAuthRequest request) {
+    log('AdaWebView:onHttpAuthRequest: host=${request.host}');
+  }
+
+  @override
+  Widget build(BuildContext context) => WebViewWidget(controller: _controller);
+
+  Future<void> _start() async {
     final metaFields = {
       ...widget.metaFields,
       'sdkType': getOsName,
@@ -215,29 +290,29 @@ class _AdaWebViewState extends State<AdaWebView> {
 
     final settingsJson = jsonEncode(settings);
 
-    await controller.evaluateJavascript(
-      source: '''
+    await _controller.runJavaScript(
+      '''
 window.adaSettings = {
   ...$settingsJson,
   lazy: true,
   parentElement: "content_frame",
+  adaReadyCallback: function(isRolledOut) {
+    onAdaReady.postMessage(JSON.stringify(isRolledOut));
+  },
   onAdaEmbedLoaded: () => {
     adaEmbed.subscribeEvent("ada:chat_frame_timeout", (data, context) => {
-       window.flutter_inappwebview.callHandler("onLoaded", data);
+      onLoaded.postMessage(typeof data === "undefined" ? "" : JSON.stringify(data));
     });
   },
   conversationEndCallback: function(event) {
-    window.flutter_inappwebview.callHandler("onConversationEnd", event);
-  },
-  adaReadyCallback: function(isRolledOut) {
-    window.flutter_inappwebview.callHandler("onAdaReady", isRolledOut);
+    onConversationEnd.postMessage(JSON.stringify(event));
   },
   toggleCallback: function(isDrawerOpen) {
-    window.flutter_inappwebview.callHandler("onDrawerToggle", isDrawerOpen);
+    onDrawerToggle.postMessage(JSON.stringify(isDrawerOpen));
   },
   eventCallbacks: {
     "*": function(event) {
-     window.flutter_inappwebview.callHandler("onEvent", JSON.stringify(event));
+      onEvent.postMessage(JSON.stringify(event));
     }
   }
 };
@@ -246,7 +321,7 @@ console.log("adaSettings: " + JSON.stringify(window.adaSettings));
     );
 
     if (widget.autostart) {
-      widget.controller?.start();
+      await widget.controller?.start();
     }
 
     Future.delayed(
@@ -259,66 +334,58 @@ console.log("adaSettings: " + JSON.stringify(window.adaSettings));
     );
   }
 
-  Future<void> _init(InAppWebViewController controller) async {
+  Future<void> _init() async {
     widget.controller?.init(
-      webViewController: controller,
+      webViewController: _controller,
       handle: widget.handle,
     );
 
-    controller.addJavaScriptHandler(
-      handlerName: 'onLoaded',
-      callback: (data) => widget.onLoaded?.call(data),
+    await _controller.addJavaScriptChannel(
+      'onLoaded',
+      onMessageReceived: (JavaScriptMessage message) {
+        final json = jsonStrToMap(message.message);
+        widget.onLoaded?.call(json);
+      },
     );
 
-    controller.addJavaScriptHandler(
-      handlerName: 'onConversationEnd',
-      callback: (event) => widget.onConversationEnd?.call(event),
+    await _controller.addJavaScriptChannel(
+      'onConversationEnd',
+      onMessageReceived: (JavaScriptMessage message) {
+        final json = jsonStrToMap(message.message);
+        widget.onConversationEnd?.call(json);
+      },
     );
 
-    controller.addJavaScriptHandler(
-      handlerName: 'onDrawerToggle',
-      callback: (isDrawerOpen) =>
-          widget.onDrawerToggle?.call(isDrawerOpen as bool),
+    await _controller.addJavaScriptChannel(
+      'onDrawerToggle',
+      onMessageReceived: (JavaScriptMessage message) {
+        final json = jsonStrToMap(message.message);
+        widget.onDrawerToggle?.call(json);
+      },
     );
 
-    controller.addJavaScriptHandler(
-      handlerName: 'onAdaReady',
-      callback: (isRolledOut) => widget.onAdaReady?.call(isRolledOut),
+    await _controller.addJavaScriptChannel(
+      'onAdaReady',
+      onMessageReceived: (JavaScriptMessage message) {
+        final json = jsonStrToMap(message.message);
+        widget.onAdaReady?.call(json);
+      },
     );
 
-    controller.addJavaScriptHandler(
-      handlerName: 'onEvent',
-      callback: (event) => widget.onEvent?.call(event),
+    await _controller.addJavaScriptChannel(
+      'onEvent',
+      onMessageReceived: (JavaScriptMessage message) {
+        final json = jsonStrToMap(message.message);
+        widget.onEvent?.call(json);
+      },
     );
   }
 
-  Future<NavigationActionPolicy?> _shouldOverrideUrlLoading(
-    InAppWebViewController controller,
-    NavigationAction navigationAction,
-  ) async {
-    final url = navigationAction.request.url;
-    debugPrint('_shouldOverrideUrlLoading: $url, host=${url?.host}');
-
-    if (url == null ||
-        url.toString() == 'about:blank' ||
-        url.toString() == widget.urlRequest?.url.toString() ||
-        url.toString().endsWith('/ada_chat_flutter/assets/embed.html') ||
-        url.host == '${widget.handle}.ada.support') {
-      return NavigationActionPolicy.ALLOW;
+  dynamic jsonStrToMap(String message) {
+    if (message.isEmpty) {
+      return <String, dynamic>{};
     }
 
-    unawaited(
-      showAdaptiveDialog(
-        context: context,
-        barrierColor: Colors.transparent,
-        builder: (context) => CustomizedWebView(
-          url: url,
-          browserSettings: widget.browserSettings,
-        ),
-        useRootNavigator: false,
-      ),
-    );
-
-    return NavigationActionPolicy.CANCEL;
+    return jsonDecode(message);
   }
 }
